# docs/swagger.yaml
definitions:
  communication.AddGameToUserRequest:
    properties:
      game_id:
        type: integer
      review:
        maxLength: 500
        type: string
      score:
        maximum: 100
        minimum: 0
        type: integer
      status:
        enum:
        - 0
        - 1
        - 2
        - 3
        type: integer
    required:
    - game_id
    type: object
  communication.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  communication.RegisterUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  communication.UpdateGameToUserRequest:
    properties:
      review:
        maxLength: 500
        type: string
      score:
        maximum: 100
        minimum: 0
        type: integer
      status:
        enum:
        - 0
        - 1
        - 2
        - 3
        type: integer
    type: object
info:
  contact: {}
  description: Checkpoint API Docs.
  title: Checkpoint API
  version: "1.0"
paths:
  /games:
    get:
      description: Get a list of games
      operationId: get-games
      parameters:
      - description: Page index
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: pageSize
        type: integer
      - description: Query for search
        in: query
        name: query
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
      summary: Get games
      tags:
      - Games
  /games/{gameId}:
    get:
      description: Get game by id
      operationId: get-game-by-id
      parameters:
      - description: Game id
        in: path
        name: gameId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
      summary: Get game by id
      tags:
      - Games
  /games/{gameId}/reviews:
    get:
      description: Get game reviews
      operationId: get-game-reviews
      parameters:
      - description: Page index
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: pageSize
        type: integer
      - description: Game id
        in: path
        name: gameId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
      summary: Get game reviews
      tags:
      - Games
  /games/rankings/top:
    get:
      description: Get games with more reviews
      operationId: get-top-games
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
      summary: Get top games
      tags:
      - Games
  /login:
    post:
      description: Login user
      operationId: login
      parameters:
      - description: User credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/communication.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: Login
      tags:
      - Authentication
  /logout:
    post:
      description: Logout user
      operationId: logout
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
      summary: Logout
      tags:
      - Authentication
  /me:
    get:
      description: Get user data from cookie token
      operationId: cookie-token-user-data
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      security:
      - cookieAuth: []
      summary: User data
      tags:
      - User
  /reviews/latest:
    get:
      description: Get Latest Reviews
      operationId: latestReviews
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
      summary: Latest Reviews
      tags:
      - Reviews
  /user/{username}/games:
    get:
      description: Get user games
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
      summary: Get user games
      tags:
      - User
  /user/{username}/games/{gameId}:
    get:
      description: Get user game by id
      parameters:
      - description: Game ID
        in: path
        name: gameId
        required: true
        type: string
      - description: Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - cookieAuth: []
      summary: Get user game
      tags:
      - User
  /user/{username}/profile:
    get:
      description: Get user profile
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
      summary: Get user profile
      tags:
      - User
  /user/games:
    post:
      description: Add a game to the user's collection
      parameters:
      - description: Game data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/communication.AddGameToUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      security:
      - cookieAuth: []
      summary: Add game to user
      tags:
      - User
  /user/games/{gameId}:
    delete:
      description: Delete user game
      parameters:
      - description: Game ID to delete
        in: path
        name: gameId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      security:
      - cookieAuth: []
      summary: Delete user game
      tags:
      - User
    put:
      description: Update user game
      parameters:
      - description: Game ID to update
        in: path
        name: gameId
        required: true
        type: string
      - description: Game data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/communication.UpdateGameToUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      security:
      - cookieAuth: []
      summary: Update user game
      tags:
      - User
  /user/profile:
    put:
      consumes:
      - multipart/form-data
      description: Update user profile details like bio, avatar etc.
      parameters:
      - description: User Bio
        in: formData
        name: bio
        type: string
      - description: User Avatar
        in: formData
        name: avatar
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - cookieAuth: []
      summary: Update user details
      tags:
      - User
  /users:
    post:
      description: Register user in the database
      operationId: register-user
      parameters:
      - description: User data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/communication.RegisterUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "409":
          description: Conflict
        "500":
          description: Internal Server Error
      summary: Register user
      tags:
      - Authentication
schemes:
- http
securityDefinitions:
  cookieAuth:
    in: cookie
    name: auth_token
    type: apiKey
swagger: "2.0"